relates to #1983: shared exception handlers to the common project
