+++ apis/sgv2-docsapi/src/test/java/io/stargate/sgv2/docsapi/api/exception/JsonParseExceptionMapperTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.stargate.sgv2.docsapi.api.common.exception;
+package io.stargate.sgv2.docsapi.api.exception;
 
 import static org.assertj.core.api.Assertions.assertThat;
