+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/exception/ConstraintViolationExceptionMapperTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package io.stargate.sgv2.docsapi.api.common.exception;
+package io.stargate.sgv2.api.common.exception;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
