+++ apis/sgv2-quarkus-common/src/test/java/io/stargate/sgv2/api/common/exception/StatusRuntimeExceptionMapperTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  *
  */
-package io.stargate.sgv2.docsapi.api.common.exception;
+package io.stargate.sgv2.api.common.exception;
 
 import static org.assertj.core.api.Assertions.assertThat;
